requirements.txt
opencv-python-headless
numpy
mediapipe
streamlit
Pillow
streamlit-webrtc